+++ controlyourtabs/keyinfo.py
@@ -19,8 +19,13 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from gi.repository import Gtk, Gdk, Gedit
-from . import log
+import gi
+gi.require_version('Gdk', '3.0')
+gi.require_version('Gtk', '3.0')
+
+from collections import namedtuple
+from gi.repository import Gdk, Gtk
+from . import editor, log
 
 
 CONTROL_MASK = Gdk.ModifierType.CONTROL_MASK
@@ -29,13 +34,37 @@ CONTROL_SHIFT_MASK = Gdk.ModifierType.CONTROL_MASK | Gdk.ModifierType.SHIFT_MASK
 
 CONTROL_KEY_LIST = [Gdk.KEY_Control_L, Gdk.KEY_Control_R] # will need to iterate through this list
 
-TAB_KEY_SET = set([Gdk.KEY_ISO_Left_Tab, Gdk.KEY_Tab])
+KEY_SETS = {
+	'tab': set([Gdk.KEY_ISO_Left_Tab, Gdk.KEY_Tab, Gdk.KEY_KP_Tab]), # what is shift numpad tab?
+	'page_up': set([Gdk.KEY_Page_Up, Gdk.KEY_KP_Page_Up]),
+	'page_down': set([Gdk.KEY_Page_Down, Gdk.KEY_KP_Page_Down]),
+	'escape': set([Gdk.KEY_Escape])
+}
 
-PAGE_KEY_SET = set([Gdk.KEY_Page_Up, Gdk.KEY_Page_Down])
+MODIFIER_KEY_SET = set(
+	[
+		Gdk.KEY_Shift_L, Gdk.KEY_Shift_R,
+		Gdk.KEY_Control_L, Gdk.KEY_Control_R,
+		Gdk.KEY_Meta_L, Gdk.KEY_Meta_R,
+		Gdk.KEY_Alt_L, Gdk.KEY_Alt_R,
+		Gdk.KEY_Super_L, Gdk.KEY_Super_R,
+		Gdk.KEY_Hyper_L, Gdk.KEY_Hyper_R,
+		Gdk.KEY_Caps_Lock, Gdk.KEY_Shift_Lock, Gdk.KEY_Num_Lock, Gdk.KEY_Scroll_Lock,
+		Gdk.KEY_ISO_Lock, Gdk.KEY_ISO_Level2_Latch,
+		Gdk.KEY_ISO_Level3_Shift, Gdk.KEY_ISO_Level3_Latch, Gdk.KEY_ISO_Level3_Lock,
+		Gdk.KEY_ISO_Level5_Shift, Gdk.KEY_ISO_Level5_Latch, Gdk.KEY_ISO_Level5_Lock,
+		Gdk.KEY_Mode_switch
+	]
+)
 
-NEXT_KEY_SET = set([Gdk.KEY_Tab, Gdk.KEY_Page_Down])
-
-ESCAPE_KEY = Gdk.KEY_Escape
+ControlKeys = namedtuple(
+	'ControlKeys',
+	[
+		*[key for key in KEY_SETS.keys()],
+		*['shift_' + key for key in KEY_SETS.keys()],
+		*[key + '_key' for key in KEY_SETS.keys()]
+	]
+)
 
 
 def default_control_held():
@@ -44,8 +73,8 @@ def default_control_held():
 def update_control_held(event, prev_statuses, new_status):
 	keyval = event.keyval
 
-	if log.query(log.INFO):
-		Gedit.debug_plugin_message(log.format("key=%s, %s, new_status=%s", Gdk.keyval_name(keyval), prev_statuses, new_status))
+	if log.query(log.DEBUG):
+		editor.debug_plugin_message(log.format("key=%s, %s, new_status=%s", Gdk.keyval_name(keyval), prev_statuses, new_status))
 
 	new_statuses = [
 		new_status if keyval == control_key else prev_status
@@ -53,7 +82,7 @@ def update_control_held(event, prev_statuses, new_status):
 	]
 
 	if log.query(log.DEBUG):
-		Gedit.debug_plugin_message(log.format("new_statuses=%s", new_statuses))
+		editor.debug_plugin_message(log.format("new_statuses=%s", new_statuses))
 
 	return new_statuses
 
@@ -61,33 +90,34 @@ def is_control_keys(event):
 	keyval = event.keyval
 	state = event.state & Gtk.accelerator_get_default_mod_mask()
 
-	if log.query(log.INFO):
-		Gedit.debug_plugin_message(log.format("key=%s, state=%s", Gdk.keyval_name(keyval), state))
+	if log.query(log.DEBUG):
+		editor.debug_plugin_message(log.format("key=%s, state=%s", Gdk.keyval_name(keyval), state))
 
 	is_control = state == CONTROL_MASK
 	is_control_shift = state == CONTROL_SHIFT_MASK
+	is_control_key = is_control or is_control_shift
 
-	is_tab = keyval in TAB_KEY_SET
-	is_page = keyval in PAGE_KEY_SET
-	is_escape = keyval == ESCAPE_KEY
+	is_key = {key: keyval in set for (key, set) in KEY_SETS.items()}
 
-	is_control_tab = (is_control or is_control_shift) and is_tab
-	is_control_page = is_control and is_page
-	is_control_escape = (is_control or is_control_shift) and is_escape
-
-	if log.query(log.DEBUG):
-		Gedit.debug_plugin_message(log.format("is_control_tab=%s, is_control_page=%s, is_control_escape=%s", is_control_tab, is_control_page, is_control_escape))
-
-	return (is_control_tab, is_control_page, is_control_escape)
-
-def is_next_key(event):
-	if log.query(log.INFO):
-		Gedit.debug_plugin_message(log.format("key=%s", Gdk.keyval_name(event.keyval)))
-
-	result = event.keyval in NEXT_KEY_SET
+	result = ControlKeys(**{
+		**{key: value and is_control for (key, value) in is_key.items()},
+		**{'shift_' + key: value and is_control_shift for (key, value) in is_key.items()},
+		**{key + '_key': value and is_control_key for (key, value) in is_key.items()}
+	})
 
 	if log.query(log.DEBUG):
-		Gedit.debug_plugin_message(log.format("result=%s", result))
+		editor.debug_plugin_message(log.format("result=%s", result))
+
+	return result
+
+def is_modifier_key(event):
+	if log.query(log.DEBUG):
+		editor.debug_plugin_message(log.format("key=%s", Gdk.keyval_name(event.keyval)))
+
+	result = event.keyval in MODIFIER_KEY_SET
+
+	if log.query(log.DEBUG):
+		editor.debug_plugin_message(log.format("result=%s", result))
 
 	return result